Comparison of a Corrective Procedure to the results of the Two One-Sided Tests (TOST) in Univariate Settings

This function renders a comparison of the alpha-TOST or the delta-TOST outputs obtained with the function `ctost` to the TOST output obtained with `tost`.

Usage

compare_to_tost(x, ticks = 30, rn = 5)

Arguments

x

A tost object, which is the output of one of the function: `ctost`.

ticks

an integer indicating the number of segments that will be printed to represent the confidence intervals.

rn

integer indicating the number of decimals places to be used (see function `round`) for the printed results.

Value

Pints a comparison between the TOST results (i.e., output of `tost`) and either the alpha-TOST or the delta-TOST results.