The delta-TOST Corrective Procedure for (Bio)Equivalence Testing

This functions is used to compute the delta-TOST, a corrective procedure of the (bio)equivalence bounds applied to the Two One-Sided Test (TOST) for (bio)equivalence testing in the univariate framework.

Usage

dtost(theta, sigma, nu, alpha, delta)

Arguments

theta

A numeric value corresponding to the difference of means.

sigma

A numeric value corresponding to the standard error.

nu

A numeric value corresponding to the number of degrees of freedom.

alpha

A numeric value specifying the significance level.

delta

A numeric value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)

Value

A tost object with the structure:

decision: A boolean variable indicating whether (bio)equivalence is accepted or not.

ci: Confidence interval at the 1 - 2*alpha level.

theta: The difference of means used for the test.

sigma: The standard error used for the test.

nu: The number of degrees of freedom used for the test.

alpha: The significance level used for the test.

delta: The (bio)equivalence limits used for the test.

method: The method used for the test (here the "delta-TOST").

Author

Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier

Examples

data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin,1,diff))/sqrt(nu)
res_dtost = cTOST:::dtost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25))
#> Error in deltahat.fun(sigma = sigma, alpha = alpha, delta = delta, nu = nu): could not find function "deltahat.fun"
compare_to_tost(res_dtost)
#> Error: object 'res_dtost' not found