Power function

Power function

Usage

power_TOST(alpha, theta, sigma_nu, nu, delta, ...)

Arguments

alpha

A numeric value specifying the significance level (default = 0.05).

theta

A numeric value corresponding to the estimated parameter of interest (such as a difference of means).

sigma_nu

A numeric value corresponding to the estimated standard error.

nu

A numeric value corresponding to the number of degrees of freedom.

delta

A numeric value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)

...

Additional parameters.

Value

The function returns a numeric value that corresponds to a probability.