Finite Sample Adjusted (Bio)Equivalence Testing

This function is used to compute finite sample corrected version of the standard (univariate or multivariate) TOST.

Usage

ctost(
  theta,
  sigma,
  nu,
  delta,
  alpha = 0.05,
  method = "optimal",
  B = 10^4,
  seed = 101010,
  correction = "offline",
  ...
)

Arguments

theta

A numeric value corresponding to the difference of means.

sigma

A numeric value corresponding to the standard error.

nu

A numeric value corresponding to the number of degrees of freedom.

delta

A numeric value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)

alpha

A numeric value specifying the significance level (default: alpha = 0.05).

method

A character value corresponding to the considered finite sample adjustment method, see Details below for more information.

B

TO DOCUMENT

seed

TO DOCUMENT

correction

TO DOCUMENT

...

Additional parameters.

Value

A tost object with the structure:

decision: A boolean variable indicating whether (bio)equivalence is accepted or not.

ci: Confidence interval at the 1 - 2*alpha level.

theta: The difference of means used for the test.

sigma: The standard error used for the test.

nu: The number of degrees of freedom used for the test.

alpha: The significance level used for the test.

delta: The (bio)equivalence limits used for the test.

method: The method used for the test (optimal-TOST, alpha-TOST and delta-TOST).

setting: The setting used (univariate or multivariate)

Details

In univariate settings, two methods are available: optimal (add ref), alpha-TOST (using method = 'alpha') and delta-TOST (using method = 'delta'). The alpha-TOST and delta-TOST are introduced in Boulaguiem et al. (2024, <https://doi.org/10.1002/sim.9993>). The former is a corrective procedure of the significance level applied to the TOST while the latter adjusts the equivalence limits. In general, the alpha-TOST appears to outperform the delta-TOST.

In multivariate setting, the only available method is the (multivariate) alpha-TOST (using method = 'alpha') introduced in Boulaguiem et al. (2024, bioarxiv).

Examples

data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = var(apply(skin,1,diff))/nu

# alpha-TOST
atost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "alpha")
atost
#> ✔ Accept (bio)equivalence
#> Equiv. Region:  |----------------0----------------|
#> Estim. Inter.:     (---------------x--------------)
#> CI =  (-0.17655 ; 0.22195)
#> 
#> Method: alpha-TOST
#> alpha = 0.05; Equiv. lim. = +/- 0.22314
#> Corrected alpha = 0.07865
#> Mean = 0.02270; Stand. dev. = 0.13428; df = 16
compare_to_tost(atost)
#> TOST:       
#> ✖ Can't accept (bio)equivalence
#> alpha-TOST: 
#> ✔ Accept (bio)equivalence
#> 
#> Equiv. Region:  |---------------0---------------|  
#> TOST:            (---------------x----------------)
#> alpha-TOST:        (-------------x--------------)  
#> 
#>                  CI - low      CI - high
#> TOST:            -0.21174       0.25715
#> alpha-TOST:      -0.17655       0.22195
#> 
#> Equiv. lim. = +/- 0.22314

# delta-TOST
dtost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "delta")
dtost
#> ✖ Can't accept (bio)equivalence
#> Corr. Equiv. Region:  |----------------0----------------|
#>       Estim. Inter.:     (--------------x---------------)
#> CI =  (-0.21174 ; 0.25715)
#> 
#> Method: delta-TOST
#> alpha = 0.05; Equiv. lim. = +/- 0.22314
#> Corrected Equiv. lim. = +/- 0.25470
#> Mean = 0.02270; Stand. dev. = 0.13428; df = 16
compare_to_tost(dtost)
#> TOST:       
#> ✖ Can't accept (bio)equivalence
#> delta-TOST: 
#> ✖ Can't accept (bio)equivalence
#> 
#> Stand. Equiv. Region:    |--------------0--------------|  
#>  Corr. Equiv. Region:  |----------------0----------------|
#>                 TOST:     (--------------x---------------)
#>           delta-TOST:     (--------------x---------------)
#> 
#>  Standard Equiv. lim. = +/- 0.22314
#> Corrected Equiv. lim. = +/- 0.25470