Two One-Sided Test (TOST) for (Bio)Equivalence Testing

This function performs a Two One-Sided Test (TOST) for (bio)equivalence testing.

tost(theta, sigma, nu, alpha, delta)

Arguments

theta

A numeric value corresponding to the difference of means (e.g. between a generic and reference drug).

sigma

A numeric value corresponding to the standard error.

nu

A numeric value corresponding to the number of degrees of freedom.

alpha

A numeric value specifying the significance level.

delta

A numeric value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)

Value

A tost object with the structure:

decision: A boolean variable indicating whether (bio)equivalence is accepted or not.

ci: Confidence interval at the 1 - 2*alpha level.

theta: The difference of means used for the test.

sigma: The standard error used for the test.

nu: The number of degrees of freedom used for the test.

alpha: The significance level used for the test.

delta: The (bio)equivalence limits used for the test.

method: The method used for the test (here the "TOST").

Author

Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier

Examples

data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin,1,diff))/sqrt(nu)
tost(theta = theta_hat, sigma = sig_hat, nu = nu,
     alpha = 0.05, delta = log(1.25))
#> ✖ Can't accept (bio)equivalence
#> Equiv. Region:  |---------------0---------------|  
#> Estim. Inter.:   (---------------x----------------)
#> CI =  (-0.21174 ; 0.25715)
#> 
#> Method: TOST
#> alpha = 0.05; Equiv. lim. = +/- 0.22314
#> Mean = 0.02270; Stand. dev. = 0.13428; df = 16